requirements.txt
cffi==1.15.1
cryptography==40.0.1
ecdsa==0.18.0
pyasn1==0.4.8
pycparser==2.21
python-jose==3.3.0
rsa==4.9
six==1.16.0
